+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - Added custom page titles and descriptions for search engine optimization
   - Moved ISSS from learn page to partners page
   - Updated link to MyApnea.Org welcome video
+  - Added didgeridoo research highlight
 - **Survey Exports**
   - Added a data dictionary export task to export all surveys into CSV format
   - Added a data export task that matches the data dictionary format and is exported into CSV format